migrate test helper functions to an abstract helper class
